+++ contracts/marketplace-contracts/contracts/installment_payment_contract/src/errors/errors.rs
@@ -1,0 +1,14 @@
+use soroban_sdk::contracterror;
+
+#[contracterror]
+#[derive(Debug, Clone, PartialEq)]
+pub enum ContractError {
+    AlreadyInstantiated = 1,
+    InvalidAmount = 2,
+    DuplicateUsers = 3,
+    ArbitratorNotAllowed = 4,
+    InvalidTimestamp = 5,
+    InvalidAgreementId = 6,
+    NotAuthorized = 7,
+    AgreementNotFOund = 8,
+}
