+++ contracts/marketplace-contracts/contracts/installment_payment_contract/src/events/events.rs
@@ -1,0 +1,1 @@
+
